requirements.txt
fastapi==0.110.1
uvicorn==0.25.0

motor==3.3.1
pymongo==4.5.0

pydantic>=2.6.4
python-dotenv>=1.0.1

pyjwt>=2.10.1
bcrypt==4.1.3
passlib>=1.7.4

email-validator>=2.2.0
tzdata>=2024.2
python-multipart>=0.0.9

qrcode>=7.4.2
